Add maxLifeTime support
This commit adds maxLifeTime to the connection pool.
Current connection lifetime is delivered from "PooledRefMetadata".
To use time information on metadata, added "SimplePoolMetricsRecorder" which
uses "Clock" to acquire current time.

Also, when "maxIdleTime=0" is set, it was evicting the connection every time,
instead of no eviction.
The behavior is fixed and unit tests are added for maxIdleTime as well.